refactor(newModal): 새음식점 추가 모달 View 로직을 파일로 분리
